requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
playwright==1.41.1
apscheduler==3.10.4
pgvector==0.2.3
sentence-transformers==2.3.1
pydantic-settings==2.0.2
beautifulsoup4==4.12.2
